serial/requirements.txt
ansitable==0.9.7
colored==1.4.4
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.3
packaging==23.1
pgraph-python==0.6.2
Pillow==9.5.0
progress==1.6
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
PyYAML==6.0
roboticstoolbox-python==1.0.3
rtb-data==1.0.0
scipy==1.10.1
six==1.16.0
spatialgeometry==1.0.3
spatialmath-python==1.0.5
swift-sim==1.0.1
websockets==11.0.2
zipp==3.15.0
